organized interface adapters by use case
