--- src/interface_adapter/SearchByIdViewModel.java
@@ -1,4 +1,0 @@
-package interface_adapter;
-
-public class SearchByIdViewModel {
-}
