+++ src/interface_adapter/search_by_id/SearchByIdState.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.search_by_id;
 
 public class SearchByIdState {
     private String recipeId = "";
