+++ src/interface_adapter/search_by_id/SearchByIdViewModel.java
@@ -1,0 +1,4 @@
+package interface_adapter.search_by_id;
+
+public class SearchByIdViewModel {
+}
